langgraph/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ironclaw"
version = "0.1.0"
description = "IronClaw — secure personal AI assistant, rewritten with LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "langchain-openai>=0.2",
    "langchain-anthropic>=0.2",
    "httpx>=0.27",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
    "asyncpg>=0.30",
    "aiofiles>=24",
    "typer>=0.13",
    "rich>=13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3",
]

[project.scripts]
ironclaw = "ironclaw.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["ironclaw*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
